server/agent/requirements.txt
livekit>=0.18.2
livekit-agents>=0.12.6
livekit-plugins-deepgram
livekit-plugins-groq
livekit-plugins-silero
livekit-plugins-turn-detector
python-dotenv
